Diploma Thesis initial upload
This is the initial upload of my Thesis.
It was written for the degree of ‚Diplom Designer‘ at the college of
fine arts ‚Hochschule für Bildende Künste Hamburg‘.
It is still in german, although i plan to translate it or certain parts
into english over time.

What still needs to be done to this website:
- clean up divs and use more relative values instead of px
- improve mobile view (font-size too small, too much width of some
elements)
- english translation